src/requirements.txt
amqp==5.0.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
decorator==4.4.2
Django==2.2.16
django-debug-toolbar==3.1.1
django-extensions==3.0.8
Faker==4.1.2
flake8==3.8.3
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
gunicorn==20.0.4
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kombu==5.0.2
lxml==4.5.2
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
traitlets==5.0.4
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==1.3.7
xlwt==1.3.0
djangorestframework
drf-yasg
